Move client related scope mutations to new file
